backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
alembic
python-dotenv
pydantic-settings
psycopg2